SOLR-13574: Fix many test and test-framework classes to not fail on After/AfterClass cleanup if assumptions fail in Before/BeforeClass setup
(cherry picked from commit 7e57d3a9d93e8acb77ce299f8c79d92df563b864)

Conflicts:
	solr/core/src/test/org/apache/solr/cloud/CleanupOldIndexTest.java

+++ solr/contrib/ltr/src/test/org/apache/solr/ltr/TestLTROnSolrCloud.java
@@ -299,7 +299,10 @@ public class TestLTROnSolrCloud extends TestRerankBase {
 
   @AfterClass
   public static void after() throws Exception {
-    FileUtils.deleteDirectory(tmpSolrHome);
+    if (null != tmpSolrHome) {
+      FileUtils.deleteDirectory(tmpSolrHome);
+      tmpSolrHome = null;
+    }
     System.clearProperty("managed.schema.mutable");
   }
